:recycle: refactor: Refactors the ChatWidget component types
