+++ apps/web/src/ui/components/chat-widget/components/svgs/index.tsx
@@ -1,9 +1,4 @@
-import type { SVGProps } from 'react'
-
-type SVGWrapperProps = SVGProps<SVGSVGElement> & {
-	title: string
-	children: SVGProps<SVGPathElement>
-}
+import type { SVGProps, SVGWrapperProps } from '../../types'
 
 function SVGWrapper({
 	children,
